iteration-11 fixing

+++ app/Http/Controllers/Admin/AdminTestController.php
@@ -60,7 +60,7 @@ class AdminTestController extends Controller
         if (isAdmin()){
                 $rhtml = ' <a  href = " '.route('onlineTestEdit',['id'=>$records->id]).'" value = "'.$records->id.'" class="btn btn-primary ViewPool" > View Test</a>';
 
-                $rhtml .= ' <i value = "'.$records->id.'" class="fas fa-trash text-danger pointer ml-3 test_id" data-target="#deletePoolModal"></i>';
+                $rhtml .= ' <i value = "'.$records->id.'" class="fas fa-trash text-danger pointer ml-3 test_id" data-toggle="modal" data-target="#deletTestModal"></i>';
             return $rhtml;
         }
       })
@@ -507,7 +507,10 @@ class AdminTestController extends Controller
             $testQuestion = TestQuestion::where('test_id', $request->id)->get();
             foreach ($testQuestion as $testQuest) {
                 $userOnlineTestAnswers = UserOnlineTestAnswers::where('question_id', $testQuest->id)->first();
-                $userOnlineTestAnswers->delete();
+                if ($userOnlineTestAnswers != null) {
+                    $userOnlineTestAnswers->delete();
+                }
+
                 $testQuest->delete();
             }
             $UserOnlineTest = UserOnlineTest::where('test_id',$OnlineTest->id)->get();
